+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@
 # 4.71.0 (2018-09-23)
 
 * Removed a bunch of help popups.
+* Added information about unique stacks.
 * Added `is:maxpower` search to return highest light items.
 * Added `is:modded` search to return items that have a mod applied.
 * Bounties with expiration times are now shown, and are sorted in front in order of expiration time.