+++ docs/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Next
 
 * Add searches `is:transmat`, `is:armormod`, `is:weaponmod`, and `is:transmat`, and removed D1 `is:primaryweaponengram`, `is:specialweaponengram`, and `is:heavyweaponengram`.
+* New option to manually sort your characters.
 
 # 4.71.0 (2018-09-23)