+++ docs/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Next
 
+# 4.72.0 (2018-09-30)
+
 * Add searches `is:transmat`, `is:armormod`, `is:weaponmod`, and `is:transmat`, and removed D1 `is:primaryweaponengram`, `is:specialweaponengram`, and `is:heavyweaponengram`.
 * Show daily gambit challenge and daily heroic adventure in milestones.